requirements.txt
requests>=2.28.0
pandas>=2.2.0
numpy>=1.25.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.12.2
python-dotenv>=1.0.0
